requirements.txt
Flask==2.2.0
Flask-RESTful==0.3.9
Flask-JWT==0.3.2
Flask-SQLAlchemy====2.5.1
Flask-Migrate==2.6.0
Flask-Script==2.0.6
alembic==1.8.1
uwsgi
psycopg2==2.9.3